change to only ever detecting a single algorithm
When looking at an existing password entry line only find the best match
This involved retooling how child classes are done and how they match a
password entry.

This simplifies the rest of the system and we don't have to deal with Array
base corner cases any more

There is the case that if a password entry is a 13 char plaintext entry it could
get mistake for a crypt entry. which i think is okay.